requirements.txt
matplotlib==3.8.0
numba==0.59.0
numpy==2.0.1
opencv_python==4.9.0.80
Pillow==10.4.0
pyamg==5.1.0
PyMatting==1.1.12
Requests==2.32.3
scikit_learn==1.2.2
scipy==1.14.0
